Revert "refactor: rename client.ts to becraft-client.ts for clarity"
Drop the client.ts → becraft-client.ts rename. The renaming of an
internal module is unrelated to the Cloudflare Workers SSR fix and
bloats the diff for this PR. The rename can be revisited as an
independent change if the ambiguity of `client.ts` becomes a real
problem.

This reverts commit b3a079d8d7d55a54f9a80b8ddef3a9db9af50e2c.
